+++ test/jest/babel.config.js
@@ -4,11 +4,8 @@ module.exports = {
     '@babel/preset-react',
   ],
   plugins: [
-    '@babel/plugin-proposal-class-properties',
-    [
-      '@babel/plugin-proposal-decorators',
-      { decoratorsBeforeExport: true },
-    ],
+    ['@babel/plugin-proposal-decorators', { legacy: true }],
+    ['@babel/plugin-proposal-class-properties', { loose: true }],
     '@babel/plugin-transform-runtime',
   ],
 };